UHF-10736: Extract IntegratinID functionality to custom service.
